Refactor search orchestration and CMU integration
